Moving the place we build the toml to the build step
Since the yarn/node artifacts are generated during the build step it makes sense to generate toml there as well.
We were orphaning the changes in the test step of the pipeline.
Also removing the paths from the toml to try and catch this sort of issue faster in the future.